requirements.txt
Django==5.0.4
django-debug-toolbar==4.3.0
psycopg[binary]==3.1.18
pydantic==2.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
split-settings==1.0.0